LRPT Decoder Prototype

+++ LRPT/Meteor/src/Display.h
@@ -19,10 +19,10 @@ private:
     uint64_t packetNumber;
     uint16_t vitErrors;
     uint16_t frameBits;
-    uint32_t rsErrors[4];
+    int32_t rsErrors[4];
     uint8_t signalQuality;
     uint8_t syncCorrelation;
-    uint8_t phaseCorrection;
+    uint16_t phaseCorrection;
     uint64_t lostPackets;
     uint16_t averageVitCorrections;
     uint8_t averageRSCorrections;
@@ -40,7 +40,7 @@ public:
     void show();
 
     void update(uint8_t scid, uint8_t vcid, uint64_t packetNumber, uint16_t vitErrors, uint16_t frameBits, int32_t *rsErrors, uint8_t signalQuality,
-            uint8_t syncCorrelation, uint8_t phaseCorrection, uint64_t lostPackets, uint16_t averageVitCorrections, uint8_t averageRSCorrections,
+            uint8_t syncCorrelation, uint16_t phaseCorrection, uint64_t lostPackets, uint16_t averageVitCorrections, uint8_t averageRSCorrections,
             uint64_t droppedPackets, int64_t *receivedPacketsPerChannel, int64_t *lostPacketsPerChannel, uint64_t totalPackets);
 };
